requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
boto3==1.38.23
botocore==1.38.23
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.2
cron-descriptor==1.4.5
distlib==0.3.9
Django==5.2.1
django-appconf==1.1.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-imagekit==5.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
filelock==3.18.0
flake8==7.2.0
google-auth==2.40.2
identify==2.6.12
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pilkit==3.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-mock==3.14.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==5.3.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
whitenoise==6.9.0
